+++ CHANGELOG.md
@@ -8,7 +8,8 @@ This page reports the main releases only and the main changes therein.
 * Implemented parallel processing for pyramid stacking algorithm
 * optimized Pyramid algorithm: now a single algorithm selects automatically the best strategy ensuring that all the processing stays approximately within a given memory budget; This avoids memory issues in case many pictures are selected. Explicit configuration is also possible for specific needs.
 * Implemented automatic subsample option for alignment, balancing and vignetting, now default.
-* Added luminosity correction in the LAB color space (8 bit only, due to cv2 limitations)
+* HLS and HSV corrections now supported for 16 bit images
+* Added luminosity correction in the LAB color space
 * Alignment module skips frames if transformation parameters are out of a reasonable ranges
 * Multilayer modules sends a warning if the estimated file size is > 1GB
 * "Run all jobs" action is enabled only if more than one job are present.